+++ libs/oef-base/src/proto_comms/ProtoMessageSender.cpp
@@ -1,25 +1,6 @@
-//------------------------------------------------------------------------------
-//
-//   Copyright 2018-2019 Fetch.AI Limited
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-//
-//------------------------------------------------------------------------------
-
-#include "oef-base/proto_comms/ProtoMessageEndpoint.hpp"
 #include "oef-base/proto_comms/ProtoMessageSender.hpp"
-
-#include "oef-messages/fetch_protobuf.hpp"
+#include "oef-base/proto_comms/ProtoMessageEndpoint.hpp"
+#include <google/protobuf/message.h>
 
 ProtoMessageSender::consumed_needed_pair ProtoMessageSender::checkForSpace(
     const mutable_buffers &data, IMessageWriter<TXType>::TXQ &txq)
@@ -47,7 +28,7 @@ ProtoMessageSender::consumed_needed_pair ProtoMessageSender::checkForSpace(
       }
       Lock lock(mutex);
 
-      uint32_t body_size = static_cast<uint32_t>(txq.front()->ByteSize());
+      uint32_t body_size = txq.front()->ByteSize();
       uint32_t head_size = sizeof(uint32_t);
       uint32_t mesg_size = body_size + head_size;
       if (chars.remainingSpace() < mesg_size)
